Implementado parte de permissões

+++ MP.ApiDotNet6.Api/Controllers/PersonController.cs
@@ -7,25 +7,33 @@ using Microsoft.AspNetCore.Mvc;
 using MP.ApiDotNet6.Application.DTOs;
 using MP.ApiDotNet6.Application.DTOs.Person;
 using MP.ApiDotNet6.Application.Services.Interfaces;
+using MP.ApiDotNet6.Domain.Authentication;
 using MP.ApiDotNet6.Domain.FiltersDb;
 
 namespace MP.ApiDotNet6.Api.Controllers
 {
-    [Authorize]
-    [Route("api/[controller]")]
-    [ApiController]
-    public class PersonController : ControllerBase
+    
+    [Route("api/[controller]")]    
+    public class PersonController : BaseController
     {
         private readonly IPersonService _personService;
-
-        public PersonController(IPersonService personService)
+        private List<string> _permissionNeeded = new List<string>() {"admin"};
+        private readonly ICurrentUser _currentUser;
+        private readonly List<string> _permissionUser;
+        public PersonController(IPersonService personService, ICurrentUser currentUser)
         {
             _personService = personService;
+            _currentUser = currentUser;
+            _permissionUser = _currentUser.Permissions.Split(",").ToList();
         }
 
         [HttpPost]
-        public async Task<ActionResult> PostAsync([FromBody] PersonDTO personDTO)
+        public async Task<IActionResult> PostAsync([FromBody] PersonDTO personDTO)
         {
+            _permissionNeeded.Add("CadastraPessoa");
+            if(!ValidPermission(_permissionUser,_permissionNeeded))
+                return Forbidden();
+
             var result = await _personService.CreateAsync(personDTO);
             if (result.IsSuccess)
                 return Ok(result);
@@ -35,8 +43,12 @@ namespace MP.ApiDotNet6.Api.Controllers
 
         [HttpGet]
         
-        public async Task<ActionResult> GetAsync()
+        public async Task<IActionResult> GetAsync()
         {
+            _permissionNeeded.Add("BuscaPessoa");
+            if(!ValidPermission(_permissionUser,_permissionNeeded))
+                return Forbidden();
+
             var result = await _personService.GetAsync();
             if (result.IsSuccess)
                 return Ok(result);
@@ -46,8 +58,12 @@ namespace MP.ApiDotNet6.Api.Controllers
 
         [HttpGet]
         [Route("paged")]
-        public async Task<ActionResult> GetPagedAsync([FromQuery] PersonFilterDb pageRequest)
+        public async Task<IActionResult> GetPagedAsync([FromQuery] PersonFilterDb pageRequest)
         {
+            _permissionNeeded.Add("BuscaPessoa");
+            if(!ValidPermission(_permissionUser,_permissionNeeded))
+                return Forbidden();
+
             var result = await _personService.GetPaged(pageRequest);
             if (result.IsSuccess)
                 return Ok(result);
@@ -57,8 +73,12 @@ namespace MP.ApiDotNet6.Api.Controllers
 
         [HttpGet]
         [Route("{id}")]
-        public async Task<ActionResult> GetIdAsync(int id)
+        public async Task<IActionResult> GetIdAsync(int id)
         {
+            _permissionNeeded.Add("BuscaPessoa");
+            if(!ValidPermission(_permissionUser,_permissionNeeded))
+                return Forbidden();
+
             var result = await _personService.GetByIdAsync(id);
             if (result.IsSuccess)   
                 return Ok(result);
@@ -67,8 +87,12 @@ namespace MP.ApiDotNet6.Api.Controllers
         }
 
         [HttpPut]
-        public async Task<ActionResult> UpdateAsync([FromBody] PersonDTO personDTO)
+        public async Task<IActionResult> UpdateAsync([FromBody] PersonDTO personDTO)
         {
+            _permissionNeeded.Add("EditaPessoa");
+            if(!ValidPermission(_permissionUser,_permissionNeeded))
+                return Forbidden();
+                
             var result = await _personService.UpdateAsync(personDTO);
             if (result.IsSuccess)
                 return Ok(result);
@@ -78,8 +102,12 @@ namespace MP.ApiDotNet6.Api.Controllers
 
         [HttpDelete]
         [Route("{id}")]
-        public async Task<ActionResult> DeleteAsync(int id)
+        public async Task<IActionResult> DeleteAsync(int id)
         {
+            _permissionNeeded.Add("DeletaPessoa");
+            if(!ValidPermission(_permissionUser,_permissionNeeded))
+                return Forbidden();
+
             var result = await _personService.RemoveAsync(id);
             if (result.IsSuccess)
                 return Ok(result);
